Adding MVC to better organice the code
