+++ src/views/tareas.pug
@@ -6,6 +6,9 @@ html
     
     h1 Lista de Tareas
     
+    if mensaje
+      p #{mensaje}
+
     table
       thead
         tr
@@ -20,14 +23,14 @@ html
             td #{tarea.nombre}
             td #{tarea.descripcion}
             td
-              form(action=`/tareas/borrar/${tarea.id}`, method="POST")
+              form(action=`/api/tareas/borrar/${tarea.id}`, method="POST")
                 input(type="submit" value="Borrar")
-              form(action=`/tareas/actualizar/${tarea.id}`, method="POST")
+              form(action=`/api/tareas/actualizar/${tarea.id}`, method="POST")
                 input(type="text", name="nombreTarea", placeholder="Nuevo nombre", required)
                 input(type="text", name="descripcionTarea", placeholder="Nueva descripción", required, maxlength="50")
                 input(type="submit" value="Editar")
     
-    form(action="/tareas/agregar", method="POST")
+    form(action="/api/tareas/agregar", method="POST")
       input(type="text", name="nombreTarea", placeholder="Nombre de la tarea", required)
       input(type="text", name="descripcionTarea", placeholder="Descripción de la tarea", required, maxlength="50")
       button(type="submit") Agregar tarea
